chore: refactor the lazy initialise class into its own file with better comments
